+++ Android/app/src/main/java/com/example/myapppersonas/repository/PersonaRepository.kt
@@ -8,7 +8,7 @@ class PersonaRepository(private val personaService: PersonaApiService) {
         val response = personaService.addPersona(persona)
         return response.isSuccessful
     }
-    suspend fun borrarPersona(id: Long): Boolean {
+    suspend fun borrarPersona(id: Int): Boolean {
         val response = personaService.deletePersona(id)
         return response.isSuccessful
     }
